% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{create_segments}
\alias{create_segments}
\title{Create phylogenetic tree segments}
\usage{
create_segments(newick_df)
}
\arguments{
\item{newick_df}{A `data.frame` containing reformatted Newick phylogenetic tree information.}
}
\value{
A `data.frame` with the coordinates of each segment on the phylogenetic tree.
}
\description{
Create phylogenetic tree segments
}
