% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{img_crop_raster}
\alias{img_crop_raster}
\title{Crop raster image}
\usage{
img_crop_raster(raster, hires_scale, tissue_positions)
}
\arguments{
\item{raster}{An image magick object that is plotted as the tissue.}

\item{hires_scale}{High resolution scale factor json file.}

\item{tissue_positions}{A `data.frame` with the pixel coordinates of each spot.}
}
\value{
A magick image object cropped to the appropriate bounding box.
}
\description{
Crop raster image
}
