% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{img_get_raster}
\alias{img_get_raster}
\title{Get Raster Image}
\usage{
img_get_raster(image_path)
}
\arguments{
\item{image_path}{File path to the underlying Visium 10X tissue image, which will be plotted.}
}
\value{
An image raster object which is used to plot the tissue image.
}
\description{
Get Raster Image
}
