% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{match_clone_barcodes}
\alias{match_clone_barcodes}
\title{Match clone names and spot barcodes}
\usage{
match_clone_barcodes(
  coordinates_df_scaled,
  clones_df,
  clone_group_name,
  coordinate_barcode_name = "cell",
  clone_barcode_name = "Barcode"
)
}
\arguments{
\item{coordinates_df_scaled}{A `data.frame` with scaled spot coordinates.}

\item{clones_df}{A `data.frame` indicating which barcodes correspond to which clones.}

\item{clone_group_name}{The name of the column that contains the clone grouping information.}

\item{coordinate_barcode_name}{The name of the column that contains the barcodes in the coordinates `data.frame`.}

\item{clone_barcode_name}{The name of the column that contains the barcodes in the clones `data.frame`.}
}
\value{
A `data.frame` with Visium 10X barcodes and their coordinates matched to clone labels.
}
\description{
Match clone names and spot barcodes
}
