% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{newick_to_graph_df}
\alias{newick_to_graph_df}
\title{Newick tree to graph data frame}
\usage{
newick_to_graph_df(newick_file)
}
\arguments{
\item{newick_file}{The path to a `.new` file containing the phylogenetic tree in Newick format.}
}
\value{
A `data.frame` containing the phylogenetic tree information re-formatted as a `data.frame`.
}
\description{
Newick tree to graph data frame
}
