% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{rotate_coordinates}
\alias{rotate_coordinates}
\title{Rotate Image Coordinates}
\usage{
rotate_coordinates(coordinates_df)
}
\arguments{
\item{coordinates_df}{A `data.frame` of spot coordinates extracted from the tissue positions file.}
}
\value{
A `data.frame` with rotated Visium 10X spot coordinates.
}
\description{
Rotate Image Coordinates
}
