% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{scale_coordinates}
\alias{scale_coordinates}
\title{Scale Coordinates}
\usage{
scale_coordinates(coordinates_df, tissue_positions)
}
\arguments{
\item{coordinates_df}{A `data.frame` with rotated spot coordinates.}

\item{tissue_positions}{A `data.frame` with tissue positions for each spot.}
}
\value{
A `data.frame` with Visium 10X coordinates scaled from 0-1.
}
\description{
Scale Coordinates
}
