% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{calculate_centroids}
\alias{calculate_centroids}
\title{Calculate centroids}
\usage{
calculate_centroids(newick_df, coordinates_df_scaled, n_repeats)
}
\arguments{
\item{newick_df}{A `data.frame` containing the reformatted Newick phylogenetic tree information.}

\item{coordinates_df_scaled}{A `data.frame` containing scaled spot coordinates and the barcodes that they correspond to.}
}
\value{
A `data.frame` with centroids calculated for each clone in the phylogenetic tree.
}
\description{
Calculate centroids
}
