% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{img_plot}
\alias{img_plot}
\title{Plot tree and spots}
\usage{
img_plot(
  raster_img,
  coordinates_df_scaled,
  newick_df,
  from_to_df,
  point_alpha = 0.8,
  hull_alpha = 0,
  hull_expansion = 0.005,
  centroid_alpha = 0.9,
  centroid_size = 8,
  segment_alpha = 0.8,
  segment_width = 2,
  segment_colour = "grey",
  fig_offset_x = 0.025,
  fig_offset_y = 0.025,
  palette
)
}
\arguments{
\item{raster_img}{The raster image object containing the tissue image to be plotted.}

\item{coordinates_df_scaled}{A `data.frame` with the scaled (0-1) spot coordinates for each Visium barcode.}

\item{newick_df}{The newick data.frame generated from the .new phylogenetic tree file.}

\item{from_to_df}{A `data.frame` containing the start and end coordinates of each branch on the phylogenetic tree.}

\item{point_alpha}{The alpha transparency value for points, default is `0.8`.}

\item{hull_alpha}{The alpha transparency value for polygons/hulls to be plotted, default is `0` (not plotted).}

\item{hull_expansion}{The hull/polygon extension amount, default is `0.005`.}

\item{centroid_alpha}{The alpha transparency value for the centroids plotted as part of the phylogenetic tree, default is `0.9`.}

\item{centroid_size}{The size of the plotted centroid points, default is `8`.}

\item{segment_alpha}{The alpha transparency value for phylogenetic tree segments, default is `0.8`.}

\item{segment_width}{The width of the phylogenetic tree segments, default is `2`.}

\item{segment_colour}{The colour of the phylogenetic tree segments, default is `"grey"`.}

\item{fig_offset_x}{The offset to adjust the histology image x location, if required. Default is `0.025`, adjustments not recommended.}

\item{fig_offset_y}{The offset to adjust the histology image y location, if required. Default is `0.025`, adjustments not recommended.}

\item{palette}{}
}
\value{
A ggplot2 object with clones, trees, and tissue image plotted together.
}
\description{
Plot tree and spots
}
