% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPhyloPlot.R
\name{SpatialPhyloPlot}
\alias{SpatialPhyloPlot}
\title{Title}
\usage{
SpatialPhyloPlot(
  visium_object,
  newick_file,
  image_file,
  clone_df,
  clone_group_column,
  clone_barcode_column = "Barcode",
  palette = "default",
  ...
)
}
\arguments{
\item{visium_object}{A `Seurat` Visium 10X object for the tumour sample of interest.}

\item{newick_file}{A file path to the phylogenetic tree in the Newick format saved as a `.new` file.}

\item{image_file}{A file path to the hires Visium 10X png image to plot.}

\item{clone_df}{A `data.frame` listing which barcode belongs to which clone.}

\item{clone_group_column}{The name of the column in `clone_df` that indicates the clones.}

\item{clone_barcode_column}{The name of the column in `clone_df` that indicates the barcode.}

\item{...}{Please see `?img_plot` for further plot parameters.}
}
\value{
A ggplot2 object with the tissue, clones, and phylogenetic trees plotted together.
}
\description{
Title
}
